third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/logical-expected.txt

This is a testharness.js-based test.
Found 220 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Can set 'margin-block' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'margin-block' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'margin-block' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'margin-block' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'margin-block' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'margin-block' to a percent: 0%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'margin-block' to a percent: -3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'margin-block' to a percent: 3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'margin-block' to a percent: calc(0% + 0%)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'margin-block' to a length: 0px
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'margin-block' to a length: -3.14em
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'margin-block' to a length: 3.14cm
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'margin-block' to a length: calc(0px + 0em)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'margin-inline' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'margin-inline' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'margin-inline' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'margin-inline' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'margin-inline' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'margin-inline' to a percent: 0%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'margin-inline' to a percent: -3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'margin-inline' to a percent: 3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'margin-inline' to a percent: calc(0% + 0%)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'margin-inline' to a length: 0px
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'margin-inline' to a length: -3.14em
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'margin-inline' to a length: 3.14cm
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'margin-inline' to a length: calc(0px + 0em)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-block-start' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'padding-block-start' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'padding-block-end' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'padding-block-end' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'padding-inline-start' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'padding-inline-start' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'padding-inline-end' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'padding-inline-end' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'padding-block' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'padding-block' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'padding-block' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'padding-block' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'padding-block' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'padding-block' to a percent: 0%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-block' to a percent: -3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-block' to a percent: 3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-block' to a percent: calc(0% + 0%)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-block' to a length: 0px
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-block' to a length: -3.14em
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-block' to a length: 3.14cm
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-block' to a length: calc(0px + 0em)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-inline' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'padding-inline' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'padding-inline' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'padding-inline' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'padding-inline' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'padding-inline' to a percent: 0%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-inline' to a percent: -3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-inline' to a percent: 3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-inline' to a percent: calc(0% + 0%)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-inline' to a length: 0px
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-inline' to a length: -3.14em
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-inline' to a length: 3.14cm
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'padding-inline' to a length: calc(0px + 0em)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-block' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'inset-block' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'inset-block' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'inset-block' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'inset-block' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'inset-block' to a percent: 0%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-block' to a percent: -3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-block' to a percent: 3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-block' to a percent: calc(0% + 0%)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-block' to a length: 0px
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-block' to a length: -3.14em
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-block' to a length: 3.14cm
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-block' to a length: calc(0px + 0em)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-inline' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'inset-inline' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'inset-inline' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'inset-inline' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'inset-inline' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'inset-inline' to a percent: 0%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-inline' to a percent: -3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-inline' to a percent: 3.14%
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-inline' to a percent: calc(0% + 0%)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-inline' to a length: 0px
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-inline' to a length: -3.14em
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-inline' to a length: 3.14cm
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'inset-inline' to a length: calc(0px + 0em)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-start' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-start' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-start' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-start' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-start' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-start' to the 'none' keyword: none
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-start-width' to the 'thin' keyword: thin
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-start-width' to the 'medium' keyword: medium
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-start-width' to the 'thick' keyword: thick
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-start-width' to a length: 0px
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-start-width' to a length: -3.14em
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-start-width' to a length: 3.14cm
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-start-width' to a length: calc(0px + 0em)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-start-color' to the 'currentcolor' keyword: currentcolor
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-start-style' to the 'none' keyword: none
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-start-style' to the 'solid' keyword: solid
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-end' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-end' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-end' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-end' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-end' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-end' to the 'none' keyword: none
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-end-width' to the 'thin' keyword: thin
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-end-width' to the 'medium' keyword: medium
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-end-width' to the 'thick' keyword: thick
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-end-width' to a length: 0px
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-end-width' to a length: -3.14em
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-end-width' to a length: 3.14cm
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-end-width' to a length: calc(0px + 0em)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-end-color' to the 'currentcolor' keyword: currentcolor
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-end-style' to the 'none' keyword: none
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-end-style' to the 'solid' keyword: solid
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-start' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-start' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-start' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-start' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-start' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-start' to the 'none' keyword: none
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-start-width' to the 'thin' keyword: thin
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-start-width' to the 'medium' keyword: medium
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-start-width' to the 'thick' keyword: thick
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-start-width' to a length: 0px
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-start-width' to a length: -3.14em
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-start-width' to a length: 3.14cm
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-start-width' to a length: calc(0px + 0em)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-start-color' to the 'currentcolor' keyword: currentcolor
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-start-style' to the 'none' keyword: none
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-start-style' to the 'solid' keyword: solid
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-end' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-end' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-end' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-end' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-end' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-end' to the 'none' keyword: none
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-end-width' to the 'thin' keyword: thin
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-end-width' to the 'medium' keyword: medium
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-end-width' to the 'thick' keyword: thick
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-end-width' to a length: 0px
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-end-width' to a length: -3.14em
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-end-width' to a length: 3.14cm
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-end-width' to a length: calc(0px + 0em)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-end-color' to the 'currentcolor' keyword: currentcolor
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-end-style' to the 'none' keyword: none
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-end-style' to the 'solid' keyword: solid
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block' to the 'none' keyword: none
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-width' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-width' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-width' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-width' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-width' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-width' to the 'thin' keyword: thin
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-width' to the 'medium' keyword: medium
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-width' to the 'thick' keyword: thick
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-width' to a length: 0px
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-width' to a length: -3.14em
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-width' to a length: 3.14cm
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-width' to a length: calc(0px + 0em)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-color' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-color' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-color' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-color' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-color' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-color' to the 'currentcolor' keyword: currentcolor
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-style' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-style' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-style' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-style' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-style' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'border-block-style' to the 'none' keyword: none
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-block-style' to the 'solid' keyword: solid
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline' to the 'none' keyword: none
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-width' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-width' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-width' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-width' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-width' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-width' to the 'thin' keyword: thin
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-width' to the 'medium' keyword: medium
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-width' to the 'thick' keyword: thick
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-width' to a length: 0px
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-width' to a length: -3.14em
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-width' to a length: 3.14cm
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-width' to a length: calc(0px + 0em)
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-color' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-color' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-color' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-color' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-color' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-color' to the 'currentcolor' keyword: currentcolor
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-style' to CSS-wide keywords: initial
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-style' to CSS-wide keywords: inherit
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-style' to CSS-wide keywords: unset
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-style' to CSS-wide keywords: revert
  assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-style' to var() references:  var(--A)
  assert_equals: expected "CSSUnparsedValue" but got "CSSStyleValue"
[FAIL] Can set 'border-inline-style' to the 'none' keyword: none
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-inline-style' to the 'solid' keyword: solid
  Failed to execute 'set' on 'StylePropertyMap': Invalid type for property
[FAIL] Can set 'border-start-start-radius' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'border-start-start-radius' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'border-start-end-radius' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'border-start-end-radius' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'border-end-start-radius' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'border-end-start-radius' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'border-end-end-radius' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'border-end-end-radius' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
Harness: the test ran to completion.
